#185 refactor Etherdelta connector add OrderBook entity
