+++ common/src/main/kotlin/fund/cyber/markets/common/model/OrderBook.kt
@@ -1,0 +1,7 @@
+package fund.cyber.markets.common.model
+
+data class OrderBook(
+    val asks: MutableList<Order>,
+    val bids: MutableList<Order>,
+    val timestamp: Long
+)
