+++ exchanges-connector/src/main/kotlin/fund/cyber/markets/connector/trade/EtherdeltaTradeConnector.kt
@@ -1,15 +1,11 @@
 package fund.cyber.markets.connector.trade
 
-import com.fasterxml.jackson.databind.JsonNode
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.module.kotlin.readValue
 import fund.cyber.markets.common.MILLIS_TO_HOURS
 import fund.cyber.markets.common.convert
 import fund.cyber.markets.common.model.TokensPair
 import fund.cyber.markets.common.model.Trade
 import fund.cyber.markets.common.model.TradeType
 import fund.cyber.markets.connector.Connector
-import fund.cyber.markets.connector.configuration.ConnectorConfiguration
 import fund.cyber.markets.connector.configuration.EXCHANGE_TAG
 import fund.cyber.markets.connector.configuration.NINE_HUNDRED_NINGTHY_FIVE_PERCENT
 import fund.cyber.markets.connector.configuration.NINGTHY_FIVE_PERCENT
@@ -17,39 +13,29 @@ import fund.cyber.markets.connector.configuration.TOKENS_PAIR_TAG
 import fund.cyber.markets.connector.configuration.TRADES_TOPIC_PREFIX
 import fund.cyber.markets.connector.configuration.TRADE_COUNT_METRIC
 import fund.cyber.markets.connector.configuration.TRADE_LATENCY_METRIC
-import fund.cyber.markets.connector.etherdelta.Erc20Contract
 import fund.cyber.markets.connector.etherdelta.EtherdeltaContract
 import fund.cyber.markets.connector.etherdelta.EtherdeltaToken
-import fund.cyber.markets.connector.etherdelta.ParityTokenRegistryContract
+import fund.cyber.markets.connector.etherdelta.EtherdeltaTokenResolver
 import io.micrometer.core.instrument.MeterRegistry
 import io.micrometer.core.instrument.Tags
 import io.micrometer.core.instrument.Timer
-import org.apache.commons.io.IOUtils
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.context.support.GenericApplicationContext
 import org.springframework.kafka.core.KafkaTemplate
 import org.springframework.stereotype.Component
 import org.web3j.protocol.Web3j
 import org.web3j.protocol.core.DefaultBlockParameter
 import org.web3j.protocol.core.methods.response.EthBlock
-import org.web3j.protocol.http.HttpService
 import org.web3j.tx.Contract
 import org.web3j.tx.ReadonlyTransactionManager
 import org.web3j.utils.Numeric
 import java.math.BigDecimal
 import java.math.BigInteger
 import java.math.RoundingMode
-import java.net.URL
-import java.nio.charset.Charset
 import java.util.*
 import java.util.concurrent.TimeUnit
-import java.util.regex.Pattern
 
-private const val ETHERDELTA_CONFIG_URL = "https://raw.githubusercontent.com/etherdelta/etherdelta.github.io/master/config/main.json"
 private const val ETHERDELTA_CONTRACT_ADDRESS = "0x8d12a197cb00d4747a1fe03395095ce2a5cc6819"
-private const val PARITY_TOKEN_REGISTRY_CONTRACT_ADDRESS = "0x5F0281910Af44bFb5fC7e86A404d0304B0e042F1"
-private const val PARITY_TOKEN_REGISTRY_EMPTY_ADDRESS = "0x0000000000000000000000000000000000000000"
 private const val ETH_SYMBOL = "ETH"
 
 @Component
@@ -58,17 +44,13 @@ class EtherdeltaTradeConnector : Connector {
 
     private val exchangeName = "ETHERDELTA"
     private val tradesTopicName by lazy { TRADES_TOPIC_PREFIX + exchangeName }
-
-    private lateinit var web3j: Web3j
-    private lateinit var exchangeTokensPairs: MutableMap<String, EtherdeltaToken>
-    private lateinit var parityTokenRegistryContract: ParityTokenRegistryContract
     private lateinit var etherdeltaContract: EtherdeltaContract
 
     @Autowired
-    private lateinit var resourceLoader: GenericApplicationContext
+    private lateinit var web3j: Web3j
 
     @Autowired
-    private lateinit var configuration: ConnectorConfiguration
+    private lateinit var etherdeltaTokenResolver: EtherdeltaTokenResolver
 
     @Autowired
     private lateinit var kafkaTemplate: KafkaTemplate<String, Any>
@@ -82,24 +64,20 @@ class EtherdeltaTradeConnector : Connector {
     override fun connect() {
         log.info("Connecting to $exchangeName exchange")
 
-        web3j = Web3j.build(HttpService(configuration.parityUrl))
-
         val gasPrice = web3j.ethGasPrice().send()
-        val tokenRegistryTransactionManager = ReadonlyTransactionManager(web3j, PARITY_TOKEN_REGISTRY_CONTRACT_ADDRESS)
         val etherdeltaTransactionManager = ReadonlyTransactionManager(web3j, ETHERDELTA_CONTRACT_ADDRESS)
-
-        parityTokenRegistryContract = ParityTokenRegistryContract.load(PARITY_TOKEN_REGISTRY_CONTRACT_ADDRESS, web3j, tokenRegistryTransactionManager, gasPrice.gasPrice, Contract.GAS_LIMIT)
         etherdeltaContract = EtherdeltaContract.load(ETHERDELTA_CONTRACT_ADDRESS, web3j, etherdeltaTransactionManager, gasPrice.gasPrice, Contract.GAS_LIMIT)
 
         log.info("Connected to $exchangeName exchange")
-
-        if (!this::exchangeTokensPairs.isInitialized) {
-            updateTokensPairs()
-        }
+        updateTokensPairs()
     }
 
     override fun disconnect() {
         log.info("Disconected from $exchangeName exchange")
+    }
+
+    override fun updateTokensPairs() {
+        etherdeltaTokenResolver.updateTokensPairs()
     }
 
     override fun isAlive(): Boolean {
@@ -140,16 +118,6 @@ class EtherdeltaTradeConnector : Connector {
                         block = web3j.ethGetBlockByHash(tradeEvent.log!!.blockHash!!, false).send()
                     }
 
-                    val getToken = exchangeTokensPairs[tradeEvent.tokenGet]
-                    val giveToken = exchangeTokensPairs[tradeEvent.tokenGive]
-
-                    if (getToken == null) {
-                        getErc20TokenDefinitonByAddress(tradeEvent.tokenGet!!)
-                    }
-                    if (giveToken == null) {
-                        getErc20TokenDefinitonByAddress(tradeEvent.tokenGive!!)
-                    }
-
                     val trade = convertTrade(tradeEvent, block)
 
                     if (trade != null) {
@@ -176,16 +144,20 @@ class EtherdeltaTradeConnector : Connector {
      * @see Trade
      */
     private fun convertTrade(tradeEvent: EtherdeltaContract.TradeEventResponse, block: EthBlock): Trade? {
-        val tokenGet = exchangeTokensPairs[tradeEvent.tokenGet]
-        val tokenGive = exchangeTokensPairs[tradeEvent.tokenGive]
+        val tokenGet = etherdeltaTokenResolver.resolveToken(tradeEvent.tokenGet)
+        val tokenGive = etherdeltaTokenResolver.resolveToken(tradeEvent.tokenGive)
 
-        val amountGet = if (tokenGet!!.base > BigInteger.ZERO) {
+        if (tokenGet == null || tokenGive == null) {
+            return null
+        }
+
+        val amountGet = if (tokenGet.base > BigInteger.ZERO) {
             BigDecimal(tradeEvent.amountGet).divide(BigDecimal(tokenGet.base))
         } else {
             BigDecimal(tradeEvent.amountGet)
         }
 
-        val amountGive = if (tokenGive!!.base > BigInteger.ZERO) {
+        val amountGive = if (tokenGive.base > BigInteger.ZERO) {
             BigDecimal(tradeEvent.amountGive).divide(BigDecimal(tokenGive.base))
         } else {
             BigDecimal(tradeEvent.amountGive)
@@ -224,136 +196,6 @@ class EtherdeltaTradeConnector : Connector {
                     price
             )
         }
-
-    }
-
-    /**
-     * Get ERC20 token definitions from different locations
-     */
-    override fun updateTokensPairs() {
-        log.info("Updating tokens pairs")
-
-        val parityTokenRegistryTokens = getParityTokenRegistryTokens()
-        val etherdeltaConfigTokens = getEtherdeltaConfigTokens()
-        val myEtherWalletTokens = getMyEtherWalletTokens()
-        exchangeTokensPairs = parityTokenRegistryTokens
-        exchangeTokensPairs.putAll(etherdeltaConfigTokens)
-        exchangeTokensPairs.putAll(myEtherWalletTokens)
-
-        log.info("Tokens pairs updated. Count: ${exchangeTokensPairs.size}")
-    }
-
-    /**
-     * Get ERC20 token definitions from etherdelta configuration JSON file.
-     * @return a map of ERC20 token address and token definition.
-     */
-    private fun getEtherdeltaConfigTokens(): MutableMap<String, EtherdeltaToken> {
-        val mapper = ObjectMapper()
-        val tokens = mutableMapOf<String, EtherdeltaToken>()
-        val base = BigInteger.TEN
-
-        val configTree =  mapper.readTree(URL(ETHERDELTA_CONFIG_URL))
-        configTree.get("tokens").asIterable().forEach { tokenNode ->
-
-            if (tokenNode.get("decimals").asInt() != 0) {
-                val tokenContractAddress = tokenNode.get("addr").asText()
-                val tokenSymbol = tokenNode.get("name").asText()
-                val tokenDecimal = tokenNode.get("decimals").asInt()
-                val tokenBase = base.pow(tokenDecimal)
-
-                tokens[tokenContractAddress] = EtherdeltaToken(tokenSymbol, tokenBase, tokenDecimal)
-            }
-        }
-
-        return tokens
-    }
-
-    /**
-     * Get ERC20 token definitions from MyEtherWallet token list.
-     * @return a map of ERC20 token address and token definition.
-     */
-    private fun getMyEtherWalletTokens(): MutableMap<String, EtherdeltaToken> {
-        val mapper = ObjectMapper()
-        val tokens = mutableMapOf<String, EtherdeltaToken>()
-        val base = BigInteger.TEN
-
-        val tokenDefinitionInputStream = resourceLoader.getResource("classpath:tokens-eth.json").inputStream
-        val tokenDefinitionString = IOUtils.toString(tokenDefinitionInputStream, Charset.forName("UTF-8"))
-
-        val tokenListTree =  mapper.readValue<Iterable<JsonNode>>(tokenDefinitionString)
-        tokenListTree.asIterable().forEach { tokenNode ->
-
-            if (tokenNode.get("decimals").asInt() != 0) {
-                val tokenContractAddress = tokenNode.get("address").asText()
-                val tokenSymbol = tokenNode.get("symbol").asText()
-                val tokenDecimal = tokenNode.get("decimals").asInt()
-                val tokenBase = base.pow(tokenDecimal)
-
-                tokens[tokenContractAddress] = EtherdeltaToken(tokenSymbol, tokenBase, tokenDecimal)
-            }
-        }
-
-        return tokens
-    }
-
-    /**
-     * Get ERC20 token definitions from parity token registry smart contract.
-     * @return a map of ERC20 token address and token definition.
-     */
-    private fun getParityTokenRegistryTokens(): MutableMap<String, EtherdeltaToken> {
-        val tokens = mutableMapOf<String, EtherdeltaToken>()
-
-        val tokensCount = parityTokenRegistryContract.tokenCount().send().toLong()
-        for (index in 0 until tokensCount) {
-            val token = parityTokenRegistryContract.token(BigInteger.valueOf(index)).send()
-
-            if (token.value1 == PARITY_TOKEN_REGISTRY_EMPTY_ADDRESS || !validBase(token.value3.toString())) {
-                continue
-            }
-
-            val tokenContractAddress = token.value1
-            val tokenSymbol = token.value2
-            val tokenBase = token.value3
-            val tokenDecimals = tokenBase.toString().length - 1
-
-            tokens[tokenContractAddress] = EtherdeltaToken(tokenSymbol, tokenBase, tokenDecimals)
-        }
-
-        return tokens
-    }
-
-    /**
-     * Get ERC20 token definitions from its own smart contract.
-     * @return a map of ERC20 token address and token definition.
-     */
-    private fun getErc20TokenDefinitonByAddress(address: String): EtherdeltaToken {
-        val transactionManager = ReadonlyTransactionManager(web3j, address)
-        val erc20Contract = Erc20Contract.load(address, web3j, transactionManager, Contract.GAS_PRICE, Contract.GAS_LIMIT)
-
-        val tokenSymbol = erc20Contract.symbol().send().trim()
-        val tokenDecimals = erc20Contract.decimals().send().intValueExact()
-        val tokenBase = BigInteger.TEN.pow(tokenDecimals)
-
-        val tokenDefinition = EtherdeltaToken(tokenSymbol, tokenBase, tokenDecimals)
-
-        log.info("Resolve new token: symbol=$tokenSymbol, decimals=$tokenDecimals")
-        exchangeTokensPairs[address] = tokenDefinition
-
-        return tokenDefinition
-    }
-
-    /**
-     * Check that @param tokenBase is valid power of 10.
-     * We need check this because parity token registry contains not valid data.
-     * @return boolean result
-     */
-    private fun validBase(tokenBase: String): Boolean {
-        val pattern = Pattern.compile("10*")
-        val matcher = pattern.matcher(tokenBase)
-
-        val result = (matcher.regionEnd() - matcher.regionStart()) == tokenBase.length || tokenBase == "0"
-
-        return result
     }
 
 }
